Fix issue# 5 (no hits found)

+++ src/TCS_MSA_PAAC.R
@@ -56,30 +56,34 @@ blastpSeq<- function(seq, start.pos = 1L, end.pos = nchar(seq),
   print(cmdblastp)
   if (silent == TRUE) system(cmdblastp, ignore.stdout = F) else system(cmdblastp)      
   #get the hit sequences Id
-  data = read.table(paste0(output.path,"out.txt"))
-  HomologousSeqIds= data$V2
-  dindex <- which(duplicated(HomologousSeqIds))
-  if(length(dindex) !=0 )
+  if(file.info(paste0(output.path,"out.txt"))$size != 0) # if there some hits are found
   {
-    HomologousSeqIds=HomologousSeqIds[-dindex]# remove duplicates if any
+    data = read.table(paste0(output.path,"out.txt"))
+    HomologousSeqIds= data$V2
+    dindex <- which(duplicated(HomologousSeqIds))
+    if(length(dindex) !=0 )
+    {
+      HomologousSeqIds=HomologousSeqIds[-dindex]# remove duplicates if any
+    }
+    # print(length(HomologousSeqIds))
+    
+    if(length(HomologousSeqIds)>=120)
+      HomologousSeqIds<-HomologousSeqIds[1:120] 
+    else
+      HomologousSeqIds<- HomologousSeqIds
+    
+    fileName<-paste0(output.path,"H.txt")
+    fileConn<-file(fileName)
+    write(as.character(HomologousSeqIds), fileConn)
+    close(fileConn)
+    
+    #get the cossponding Fasta file
+    getseqcmd= paste0(shQuote(Sys.which('blastdbcmd')),' -db ',shQuote(database.path), ' -entry_batch ', fileName, ' -out ', paste0(output.path,"/seq.txt"))
+    # print(getseqcmd)
+    if (silent == TRUE) system(getseqcmd, ignore.stdout = TRUE) else system(getseqcmd)
+  }else{
+    print (paste0( output.path,"---No hits found"))
   }
-  # print(length(HomologousSeqIds))
-  
-  if(length(HomologousSeqIds)>=120)
-    HomologousSeqIds<-HomologousSeqIds[1:120] 
-  else
-    HomologousSeqIds<- HomologousSeqIds
-  
-  fileName<-paste0(output.path,"H.txt")
-  fileConn<-file(fileName)
-  write(as.character(HomologousSeqIds), fileConn)
-  close(fileConn)
-  
-  #get the cossponding Fasta file
-  getseqcmd= paste0(shQuote(Sys.which('blastdbcmd')),' -db ',shQuote(database.path), ' -entry_batch ', fileName, ' -out ', paste0(output.path,"/seq.txt"))
-  # print(getseqcmd)
-  if (silent == TRUE) system(getseqcmd, ignore.stdout = TRUE) else system(getseqcmd)
-  
 }
 
 FilteredMSA= function(path)
@@ -87,34 +91,35 @@ FilteredMSA= function(path)
   #print(path)
   
   setwd(path)
-  
-  tcsScorecmd<-paste0("t_coffee seq.txt -mode psicoffee -blast_server=LOCAL -protein_db ",dbpath,"uniref50-tm.fasta  -output tcs_residue_filter3_fasta,clustalw_aln,tcs_column_filter4_fasta,score_html")
+  if(file.exists("seq.txt"))
+  {
+  tcsScorecmd<-paste0("t_coffee seq.txt -mode psicoffee -blast_server=LOCAL -protein_db ",dbpath,"uniref50-tm.fasta -output tcs_residue_filter3_fasta,clustalw_aln,tcs_column_filter4_fasta,score_html")
   system(tcsScorecmd)
   system('rm *.prf')
   #Removing columns with 80%gaps or more 
- # print("4-Removing columns with 80%gaps or more ")
+  # print("4-Removing columns with 80%gaps or more ")
   removegapscmd= paste0("t_coffee -other_pg seq_reformat -in seq.tcs_column_filter4_fasta", " -output  fasta > filteredSeq.fasta")
   system(removegapscmd)
-  
+  }
 }
 
 PreparedataforMSAAAC = function(seq, start.pos = 1L, end.pos = nchar(seq), 
                                 blastp.path = NULL, makeblastdb.path = NULL, 
                                 database.path = NULL, iter = 5, silent = TRUE, 
                                 evalue = 10L, output.path=resultspath) {
-    #1- run blastp on datafiles
+  #1- run blastp on datafiles
   seqname=names(seq)
   SeqDirectory=paste0(output.path,names(seq),"/")
   if((! dir.exists(SeqDirectory)) ) #first time looking at this sequence
   {
-  dir.create(SeqDirectory, showWarnings = FALSE, recursive = FALSE, mode = "0777")   
-  blastpSeq(seq, start.pos , end.pos ,  blastp.path, makeblastdb.path , database.path , silent ,evalue, SeqDirectory)
+    dir.create(SeqDirectory, showWarnings = FALSE, recursive = FALSE, mode = "0777")   
+    blastpSeq(seq, start.pos , end.pos ,  blastp.path, makeblastdb.path , database.path , silent ,evalue, SeqDirectory)
   }
   
   #2- Do Filtered MSA 
   if((! file.exists(paste0(SeqDirectory, "filteredSeq.fasta"))) || (file.info(paste0(SeqDirectory, "filteredSeq.fasta"))$size  ==0))
-  FilteredMSA(SeqDirectory)  
-  }
+    {FilteredMSA(SeqDirectory)}
+}
 
 
 
@@ -136,22 +141,20 @@ MSA_TCS_PAAC<- function(subset,fastafile)
   # step#2 analyse data ( collect blast hits, do MSA, remove low socore col)
   
   dfMSADC <- data.frame(matrix(ncol = 400+1, nrow =0)) 
-  
-  
+   
   setwd(paste0(intermediateFiles))
   AAfiles<-  names(seqs)
-  
-  
-  
   ListofSubsetDetailedMSADCStatistics <- matrix(ncol=400+1,nrow=length(AAfiles))
-  
-  rownames( ListofSubsetDetailedMSADCStatistics )<- AAfiles
+   rownames( ListofSubsetDetailedMSADCStatistics )<- AAfiles
   print("=====================================================================================")
   print(subset)
   for(i in c(1:length(AAfiles)))
   {
     subdirName<- paste0(getwd(),"/",AAfiles[i],"/")
     print(subdirName)
+    
+    if((file.exists(paste0(subdirName, "filteredSeq.fasta"))) && (file.info(paste0(subdirName, "filteredSeq.fasta"))$size  >0))
+    {
     se<-  readFASTA(paste0(subdirName, "filteredSeq.fasta"))
     seqlist<- unlist(lapply(se,as.character))
     seqlist<-sapply(seqlist,gsub,pattern="[^GPAVLIMCFYWHKRQNEDST]",replacement="") # getting rid of gaps? needs more work
@@ -164,6 +167,11 @@ MSA_TCS_PAAC<- function(subset,fastafile)
     #colnames(outputDC)<- names(MSADC[[1]])  
     DC<- apply(outputDC, 2, mean)
     ListofSubsetDetailedMSADCStatistics[i,]<-   c("NA",DC)
+    }else{ # if the sequence does not have any hits, just compute the PAAC from the single query
+      print("filteredSeq.fasta does not exist")
+      DC<- extractDC(as.character(seqs[i]))
+      ListofSubsetDetailedMSADCStatistics[i,]<-   c("NA",DC)
+    }
   }
   dfMSADC<- rbind(dfMSADC,ListofSubsetDetailedMSADCStatistics)
